+++ Modules/OfficeOnlineServerDsc/Examples/NewFarm.ps1
@@ -1,6 +1,6 @@
 configuration WebAppsFarm
 {
-    Import-DscResource -ModuleName xOfficeOnlineServer
+    Import-DscResource -ModuleName OfficeOnlineServerDsc
 
     node localhost
     {