+++ Tests/Integration/MSFT_OfficeOnlineServerInstall.config.ps1
@@ -12,7 +12,7 @@
 
 
 configuration MSFT_xOfficeOnlineServerInstall_config {
-    Import-DscResource -ModuleName 'xOfficeOnlineServer'
+    Import-DscResource -ModuleName 'OfficeOnlineServerDsc'
     node localhost {
         # TODO: Modify ResourceFriendlyName (e.g. xFirewall Integration_Test)
         xOfficeOnlineServerInstall Integration_Test {