+++ Tests/Integration/MSFT_OfficeOnlineServerWebAppsFarm.Integration.tests.ps1
@@ -13,7 +13,7 @@
 #>
 
 # TODO: Customize these parameters...
-$Global:DSCModuleName      = 'xOfficeOnlineServer' # Example xNetworking
+$Global:DSCModuleName      = 'OfficeOnlineServerDsc' # Example xNetworking
 $Global:DSCResourceName    = 'MSFT_xOfficeOnlineServerWebAppsFarm' # Example MSFT_xFirewall
 # /TODO